Added x,y,z scaling of keyboards
You can now grow or shrink the size of the keyboards, to grow press
i,o,p for x,y,z and to shrink press j,k,l for x,y,z. The text to voice
is acting a little funny but for the most part everything works.
Another to add is a forward/reverse switch keyboard.

Signed-off-by: John Morio Sakaguchi <morio.earth@gmail.com>