Added a fourth keyboard, direction toggle
The fourth keyboard toggles the typewriter key select direction to the
opposite direction. So keyboard forward goes backwards and the
keyboard backwards goes forward when the fourth keyboard is activated.
Also added messages for all adjustments to a keyboard e.g. scale_x,
rotate_y, translate_z, etc. Shift "+" and shift "-" adjust the scale
of the adjustments by a factor of 10.